Add private Task.StateFlags property to help with debugging (#55297)
It's a pain when you want to debug something with tasks, you look at its m_stateFlags, and you're met with an integer you have to decode by looking at consts defined in Task.  This changes those consts to be an enum and adds a private property that returns the flags as that enum, to help with debugging.