Require CFB8 for persisted CNG keys in CFB mode
This changes AesCng and TripleDESCng to require the feedback size to be
set to '8' when in CFB mode and using a persisted CNG key.

Prior to this change, BasicSymmetricCipherNCrypt ignored the feedback
size which resulted in CFB8 always being used, even if the FeedbackSize
was set to another value. However, when padding was applied, the padding
size would be padded to the feedback size. In the case of AesCng, this
would mean it would be encrypted with CFB8 and padded as if it were
CFB128.

This changes the implementation so that the feedback size is required
to be set to 8 for persisted keys. No change is made for ephemeral keys.